Revert "Completed task 3"
