--- image.md
@@ -1,15 +1,0 @@
-## Task3 Screenshots
-
-### myMovies collection creation successful
-![mongo and mongod](/images/myMoviesCollectionCreationSuccessful.jpg)
-
-
-### Quering movie collection
-![mongo and mongod](/images/quering-movie-collection.jpg)
-
-### update movie collection
-![mongo and mongod](/images/update-movie-collection.jpg)
-
-### current documents in movie collection
-![mongo and mongod](/images/current-documents.jpg)
-
